// src/app/shared/types/routes.ts
export enum MainRoutes {
  DEFAULT = "",
  NOT_FOUND = "**",
  ADMIN = "admin",
  BORROWER = "borrower",
  FO = "fo",
  LOGIN = "login",
  STORE = "store",
}

export enum AdditionalRoutes {
  SETTINGS = "settings",
  SUPPORT = "support",
  STATISTICS = "statistics",
  BORROWERS_LIST = "borrowers",
  FO_LIST = "fo",
  STORES_LIST = "stores",
  SURVEY_SETTINGS = "survey-settings",
  CALCULATOR = "calculator",
  TICKETS = "tickets",
}